CLOUD-53510 create factory for container constraint creation
