+++ core/src/main/java/com/sequenceiq/cloudbreak/service/cluster/ContainerService.java
@@ -1,14 +1,13 @@
 package com.sequenceiq.cloudbreak.service.cluster;
 
-import java.util.List;
+import com.sequenceiq.cloudbreak.domain.Container;
+import com.sequenceiq.cloudbreak.repository.ContainerRepository;
+import org.springframework.stereotype.Service;
 
 import javax.inject.Inject;
 import javax.transaction.Transactional;
-
-import org.springframework.stereotype.Service;
-
-import com.sequenceiq.cloudbreak.domain.Container;
-import com.sequenceiq.cloudbreak.repository.ContainerRepository;
+import java.util.List;
+import java.util.Set;
 
 @Service
 @Transactional
@@ -20,5 +19,10 @@ public class ContainerService {
     @Transactional(javax.transaction.Transactional.TxType.NEVER)
     public Iterable<Container> save(List<Container> containers) {
         return containerRepository.save(containers);
+    }
+
+    @Transactional(javax.transaction.Transactional.TxType.NEVER)
+    public Set<Container> findContainersInCluster(Long clusterId) {
+        return containerRepository.findContainersInCluster(clusterId);
     }
 }
